ansible/roles/app/files/app/requirements.txt
Flask
psycopg2-binary
prometheus_flask_exporter
